chore(http-problem): set MSRV to 1.60
I introduced `http_problem` to a friend of mine and noticed that we require the
latest nightly version with an unstable feature that's not actually used. I then
ran `cargo-msrv` and found out that the MSRV is actually `stable 1.60`.

We may even be able to bump it down to something like 1.58, but I didn't manage
to do it because of tokio-postgres, but there might be a way.

How do you feel about this @luisholanda @lffg?

GitOrigin-RevId: 2799dad730914bb631af1a60ca713aeadc67415e