feat(http-problem): add unprocessable not found
It is not uncommon to find parts of the codebase where we rely on an `Option`
being `Some` to finish some processing. As returning a "not found" with a 404
error may be misleading, I propose this helper function that allows an
`unprocessable` error to be returned in a less verbose way.

GitOrigin-RevId: 823b558233c846c1b68932a7b5170734665e456e